pyproject.toml
[project]
name = "geography-geyser-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "cloudinary>=1.44.1",
    "cryptography>=46.0.3",
    "dj-rest-auth>=7.0.1",
    "django-allauth>=65.13.1",
    "django-cloudinary-storage>=0.3.0",
    "django-environ>=0.12.0",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt>=5.5.1",
    "pillow>=12.0.0",
    "psycopg2-binary>=2.9.11",
    "requests>=2.32.5",
]
